+++ app/src/main/java/woowacourse/movie/view/alarm/ReservationNotificationManager.kt
@@ -9,8 +9,8 @@ import androidx.core.app.ActivityCompat
 import androidx.core.app.NotificationCompat
 import androidx.core.app.NotificationManagerCompat
 import woowacourse.movie.R
-import woowacourse.movie.view.ReservationCompletedActivity
 import woowacourse.movie.view.model.ReservationUiModel
+import woowacourse.movie.view.reservationcompleted.ReservationCompletedActivity
 
 class ReservationNotificationManager(
     private val context: Context
