+++ app/src/main/java/woowacourse/movie/view/reservation/ReservationActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.reservation
 
 import android.content.Context
 import android.content.Intent
