+++ app/src/main/java/woowacourse/movie/view/reservationcompleted/ReservationCompletedActivity.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.view
+package woowacourse.movie.view.reservationcompleted
 
 import android.content.Context
 import android.content.Intent
