+++ app/src/main/res/layout/activity_reservation_completed.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.ReservationCompletedActivity">
+    tools:context=".view.reservationcompleted.ReservationCompletedActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
